third_party/blink/web_tests/wpt_internal/prerender/restriction-element-request-fullscreen.html

<!DOCTYPE html>
<title>Prerendering cannot invoke Element.requestFullscreen</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // Start prerendering a page that attempts to invoke
  // element.requestFullscreen.
  // This API is activation-gated so it's expected to fail.
  startPrerendering(`resources/request-fullscreen.html`);

  const result = await gotMessage;
  assert_equals(result, 'request failed');
}, 'prerendering page cannot invoke element.requestFullscreen');

</script>
</body>
